requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
firebase-admin==6.3.0
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1
numpy==1.26.2
pandas==2.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
aiohttp==3.9.1
requests==2.31.0
pydantic-settings==2.1.0